Support signed integer parameters without units
The zero offset is changed to be non-optional because it didn't actually
provide extra information or different behaviour. The simple and range
kinds are merged to avoid having two different code paths for the zero
offset.

The copy-paste error where + was always shown in front of positive unit
values, not just for ranges where the signs differ at the end, is fixed.

Relevant parameters now have signed definitions.