Make has_drum_key a property of the parameter, not the block
It slipped my mind before that this can't be a block-level thing because
DRUM MAP NAME isn't per-key.

This also changes the output to reflect this hierarchy. I was trying to
avoid this before because I thought it was a weird ordering, but it
matches the ordering of the bytes in the SysEx message, so maybe it's
actually better this way. It's also nice how now there isn't an
intermediate shade of gray for the drum key number, it's actually a
darker shade of grey than the normal parameter one.